refactor: extract workspace repository lifecycle
